refactor: Extract CheckDMGMChannelPermissions to separate file
- Move CheckDMGMChannelPermissions function to permissions_dmgm.go
- Refactor switch statement to reduce code duplication
- Use variable to determine required permission before checking
- Improves maintainability and reduces merge conflicts

Addresses review feedback from PR #114
